refactor: separate function registratipn
